+++ tests/circuit_breaker/test_sync.py
@@ -309,6 +309,26 @@ class TestCircuitBreaker:
 
         assert cache.get(failure_cache_key) == 0
 
+    def test_should_create_failure_cache_when_no_request_cache_key(
+        self,
+        cache,
+        should_not_open_rule_without_request_cache_key,
+        failure_cache_key,
+        request_cache_key,
+    ):
+        assert cache.get(failure_cache_key) is None
+
+        with pytest.raises(ValueError):
+            with CircuitBreaker(
+                rule=should_not_open_rule_without_request_cache_key,
+                cache=cache,
+                failure_exception=MyException,
+                catch_exceptions=(ValueError,),
+            ):
+                fail_function()
+
+        assert cache.get(failure_cache_key) == 1
+
     def test_should_call_expire_if_incr_returns_one(
         self,
         cache,